+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 4.18.3 (October 31, 2024)
+
 ### Fixed
 
 - Objects created on clusters older than 1.18 will no longer see a